ci: fix main tests (#14)
* ci: update chezmoi source file

* perf(mise): don’t quiet mise install command

* perd: limit parallel of mise install command

* fix(mise): invalid mise job option

* chore: disable install rust tool by default in mise

* chore: add logs when start initiate zinit

* chore: add more logs when run mise install tools

* fix(mise): rollback [env] field try to fix the deadlock

https://github.com/jdx/mise/discussions/3548

* revert: bring rust back

* chore: try alternative way to run mise install

* chore: add verbose when run mise install

* chore: try to verbose mise commands

* chore: debug even more

* fix(mise): remove npm from global mise tools

* perf(mise): remove copier from global mise tools

* ci: support custom chezmoi arguments when run main tests


* ci: use --force on chezmoi to always say yes on prompt


* revert: mise is now install in parallel


* perf(script): improve setup-* and cleanup-* script


* fix: cannot use chezmoi if we didn't install it :sob:


* fix(mise): wrong tool name to install


* fix(script): try to activate mise on setup-chezmoi.sh


* fix(script): mise activate should be bash on bash script


* feat(script): add verify-chezmoi.sh to check the init & apply results


* ci: update verify-chezmoi should run with secrets


* ci: add 1p token to setup-secrets.sh script


* ci: wrong condition


* chore(script): add similar banner on each verify


* chore: reduce logs from run script and add mise list on verify command


---------

+++ home/dot_local/bin/executable_setup-secrets.sh
@@ -6,61 +6,85 @@
 ##   - GITHUB_TOKEN
 ##   - GPG_FINGERPRINT
 
-set -e
-
 export OP_ACCOUNT="my.1password.com"
 export OP_VAULT="${OP_VAULT:-64yg7lvccdzebup52w2nmzoady}"
 
 setup_1password() {
-  : "${OP_SERVICE_ACCOUNT_TOKEN:=${OP_TOKEN:-${TOKEN}}}"
   export OP_SERVICE_ACCOUNT_TOKEN
-
   if [ -z "$OP_SERVICE_ACCOUNT_TOKEN" ]; then
-    echo -n 'Enter 1Password service account token: '
-    read -rs OP_SERVICE_ACCOUNT_TOKEN
-    echo
+    if [ -n "$CI" ]; then
+      printf 'Missing 1Password token, please set OP_SERVICE_ACCOUNT_TOKEN environment variable' >&2
+      return 2
+    elif [ -n "$DOCKER" ]; then
+      echo -n 'Enter 1Password service account token: '
+      read -rs OP_SERVICE_ACCOUNT_TOKEN
+      echo
+    else
+      op signin
+    fi
   fi
 
   ## Check is signed in
-  op whoami
+  if ! op whoami; then
+    return 5
+  fi
 }
 
 setup_github() {
+  export GITHUB_TOKEN
+
   if [ -z "$GITHUB_TOKEN" ]; then
     export TOKEN="op://$OP_VAULT/GITHUB_TOKEN/password"
     op run --no-masking -- printenv TOKEN | gh auth login --with-token
     unset TOKEN
 
     GITHUB_TOKEN="$(gh auth token)"
-    export GITHUB_TOKEN
   fi
 
   ## Check is signed in
-  gh auth status
+  if ! gh auth status --active; then
+    return 5
+  fi
 }
 
 setup_gpg() {
-  local gpg_key_op_item="fp3xx4qmu7ocxr2r546q6deocu"
+  if [ -z "$GPG_FINGERPRINT" ]; then
+    local gpg_key_op_item="fp3xx4qmu7ocxr2r546q6deocu"
 
-  local gpg_key_path gpg_finger_path
-  gpg_key_path="$(mktemp)"
-  gpg_finger_path="$(mktemp)"
+    local gpg_key_path gpg_finger_path
+    gpg_key_path="$(mktemp)"
+    gpg_finger_path="$(mktemp)"
 
-  op item get "$gpg_key_op_item" \
-    --vault "$OP_VAULT" \
-    --field label=Fingerprint | tee "$gpg_finger_path"
-  op document get "$gpg_key_op_item" \
-    --vault "$OP_VAULT" \
-    --out-file "$gpg_key_path" >/dev/null
-  gpg --quiet --import "$gpg_key_path"
+    op item get "$gpg_key_op_item" \
+      --vault "$OP_VAULT" \
+      --field label=Fingerprint | tee "$gpg_finger_path"
+    op document get "$gpg_key_op_item" \
+      --vault "$OP_VAULT" \
+      --out-file "$gpg_key_path" --force
+    gpg --quiet --import "$gpg_key_path"
 
-  GPG_FINGERPRINT="$(cat "$gpg_finger_path")"
-  export GPG_FINGERPRINT
+    GPG_FINGERPRINT="$(cat "$gpg_finger_path")"
+    export GPG_FINGERPRINT
 
-  rm "$gpg_key_path" "$gpg_finger_path"
+    rm "$gpg_key_path" "$gpg_finger_path"
+  fi
+
+  if ! gpg --list-secret-keys "$GPG_FINGERPRINT"; then
+    return 5
+  fi
 }
 
+printf "\n==============================================
+Set up %s
+==============================================\n" "1Password"
 setup_1password
 
+printf "\n==============================================
+Set up %s
+==============================================\n" "GitHub Authentication"
 setup_github
+
+printf "\n==============================================
+Set up %s
+==============================================\n" "GPG key"
 setup_gpg
